backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlmodel==0.0.8
python-multipart==0.0.5
tensorflow==2.15.0
pillow==9.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
